+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- authenticate client
+- authenticateRequest middleware
+- authenticateRequest middleware to the affiliate route
+
 ## [0.12.0] - 2021-12-08
 
 ### Changed